+++ packages/core/src/index.ts
@@ -4,7 +4,9 @@ export {
   SingletonFactory,
   TransientFactory,
   OneShotFactory,
+  type DependencyKey,
 } from './serviceFactory.js'
-export { ServiceKey, SelectorKey, ServiceSelectorKey } from './serviceKey.js'
+export { ServiceKey } from './serviceKey.js'
+export { SelectorKey, ServiceSelectorKey } from './selectorKey.js'
 export { type Selector, type ServiceSelector } from './selector.js'
 export * from './errors.js'